Move About us, Contact, Blog and Privacy policy pages to app folder (#153)
* Move placeholder pages to app folder

* update new next/image component

* small cleanup

---------
