+++ app/blog/page.tsx
@@ -1,7 +1,5 @@
 import dayjs from 'dayjs';
 
-import Head from '~/components/globals/Head';
-
 import keywords from '~/data/keywords.json';
 import Container from '~/layouts/Container';
 
@@ -42,56 +40,52 @@ const articles = [
   }
 ];
 
-const Blog = () => {
-  return (
-    <>
-      <Head
-        title="SquareOne - Blog"
-        description="Sit excepteur proident est commodo laboris consectetur ea tempor officia."
-        keywords={keywords.blog}
-        url={`${NEXT_PUBLIC_BASE_URL}/blog`}
-      />
-
-      <Container className="py-20">
-        <h1 className="text-align pb-20 text-4xl font-libre uppercase">Blog</h1>
-        <ul>
-          {articles.map((article, i) => (
-            <li key={i} className="border border-gray-medium mb-8 py-4 px-10">
-              <p className="mb-3 flex items-center font-quicksand">
-                <svg
-                  className="w-4 h-4 mr-2"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
-                  ></path>
-                </svg>
-                Published: {dayjs(article.created_at).locale('en').format('DD MMMM YYYY')}
-              </p>
-              <a>
-                <h2 className="hover:underline font-libre text-headline-3 2xl:text-[2rem] mb-4 cursor-pointer">
-                  {article.title}
-                </h2>
-              </a>
-              <p className="mb-4 flex items-center font-quicksand">{article.description}</p>
-              <a
-                rel="noopener noreferrer"
-                className="w-fit flex items-center font-semibold min-w-fit mt-[1.75rem]"
-              >
-                <span className="mr-2 cursor-pointer underline font-quicksand">View more</span>
-              </a>
-            </li>
-          ))}
-        </ul>
-      </Container>
-    </>
-  );
+export const metadata = {
+  title: 'SquareOne - Blog',
+  description: 'Sit excepteur proident est commodo laboris consectetur ea tempor officia.',
+  keywords: keywords.blog,
+  url: `${NEXT_PUBLIC_BASE_URL}/blog`
 };
+
+const Blog = () => (
+  <Container className="py-20">
+    <h1 className="text-align pb-20 text-4xl font-libre uppercase">Blog</h1>
+    <ul>
+      {articles.map((article, i) => (
+        <li key={i} className="border border-gray-medium mb-8 py-4 px-10">
+          <p className="mb-3 flex items-center font-quicksand">
+            <svg
+              className="w-4 h-4 mr-2"
+              fill="none"
+              stroke="currentColor"
+              viewBox="0 0 24 24"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth="2"
+                d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
+              ></path>
+            </svg>
+            Published: {dayjs(article.created_at).locale('en').format('DD MMMM YYYY')}
+          </p>
+          <a>
+            <h2 className="hover:underline font-libre text-headline-3 2xl:text-[2rem] mb-4 cursor-pointer">
+              {article.title}
+            </h2>
+          </a>
+          <p className="mb-4 flex items-center font-quicksand">{article.description}</p>
+          <a
+            rel="noopener noreferrer"
+            className="w-fit flex items-center font-semibold min-w-fit mt-[1.75rem]"
+          >
+            <span className="mr-2 cursor-pointer underline font-quicksand">View more</span>
+          </a>
+        </li>
+      ))}
+    </ul>
+  </Container>
+);
 
 export default Blog;
